% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit data to linear or quadratic model}
\usage{
fit_model(df, degree = 1, temp_correct = FALSE, exclusion = list())
}
\arguments{
\item{df}{A dataframe containing species name, mass, metabolic rates, body temperature}

\item{degree}{An integer which is the degree of polynomial: 1 (default)}

\item{temp_correct}{A flag for temperature correction, TRUE or FALSE (default)}

\item{exclusion}{A list of species name to exclude}
}
\value{
The coefficients for the fitted model
}
\description{
Fit data to linear or quadratic model
}
