% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load raw data}
\usage{
load_data(filename)
}
\arguments{
\item{filename}{}
}
\value{

}
\description{
This turns the raw data Excel file into a dataframe.
}
\examples{
load_data("rawdata.xls")
}
